requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
attrs==25.3.0
certifi==2025.8.3
dotenv==0.9.9
frozenlist==1.7.0
idna==3.10
iso8601==2.1.0
magic-filter==1.0.12
multidict==6.6.4
propcache==0.3.2
pydantic==2.11.9
pydantic_core==2.33.2
pypika-tortoise==0.6.2
python-dotenv==1.1.1
pytz==2025.2
tortoise-orm==0.25.1
typing-inspection==0.4.1
typing_extensions==4.15.0
yarl==1.20.1
